refactor(oss): streamline request header handling and improve utility functions
