+++ core/src/oss/utils.rs
@@ -1,12 +1,17 @@
 use crate::error::Error;
+use crate::oss::Client;
+use crate::oss::object::utils::partition_header;
+use crate::oss::sign_v4::{HTTPVerb, SignV4Param};
 use base64::{Engine, engine::general_purpose};
+use common_lib::helper::gmt_format;
 use md5::{Digest, Md5};
 use regex::Regex;
 use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
 use serde::Serialize;
-use std::collections::HashMap;
+use std::collections::{BTreeMap, BTreeSet, HashMap};
 use std::path::Path;
 use tokio::io::AsyncReadExt;
+use url::Url;
 
 pub fn get_content_md5(bytes: &[u8]) -> String {
     use md5::{Digest, Md5};
@@ -97,4 +102,43 @@ fn validate_object_name_test() {
     assert!(!is_valid_object_name("/"));
     assert!(!is_valid_object_name("//foo"));
     assert!(!is_valid_object_name("/foo//bar"));
+}
+
+pub(crate) fn get_request_header(
+    client: &Client,
+    req_header_map: HashMap<String, String>,
+    request_url: &Url,
+) -> HeaderMap {
+    // 把需要签名的header和不需要签名的header分开
+    let (sign_map, remaining_map) = partition_header(req_header_map);
+
+    // 创建CanonicalHeaders，把所有需要签名的header放到CanonicalHeader中
+    let mut canonical_header = BTreeMap::new();
+    canonical_header.extend(sign_map.iter().map(|(k, v)| (k.as_str(), v.as_str())));
+
+    canonical_header.insert("x-oss-content-sha256", "UNSIGNED-PAYLOAD");
+    canonical_header.insert("host", request_url.host_str().unwrap());
+
+    let mut additional_header = BTreeSet::new();
+    additional_header.insert("host");
+    let now = time::OffsetDateTime::now_utc();
+    let sign_v4_param = SignV4Param {
+        signing_region: &client.region,
+        http_verb: HTTPVerb::Put,
+        uri: request_url,
+        bucket: Some(&client.bucket),
+        header_map: &canonical_header,
+        additional_header: Some(&additional_header),
+        date_time: &now,
+    };
+    let authorization = client.sign_v4(sign_v4_param);
+
+    // 把canonical_header转化为最终的header，补齐剩下的未参与签名计算的header
+    // 包括：剩下必要的公共请求头，api header中的非签名字段
+    let mut header = canonical_header.into_iter().collect::<HashMap<_, _>>();
+    header.insert("Authorization", &authorization);
+    let gmt = gmt_format(&now);
+    header.insert("Date", &gmt);
+    header.extend(remaining_map.iter().map(|(k, v)| (k.as_str(), v.as_str())));
+    into_request_header(header)
 }
